+++ tests/expr/bugs/do2.c
@@ -1,11 +1,11 @@
 #define INTRINSICS_IMPLEMENTATION
 #include "intrinsics.h"
 
-i64 expr(void) {
-i64 result_; {
-    i64 inner_before_ = 0;
+i32 expr(void) {
+i32 result_; {
+    i32 inner_before_ = 0;
 
-    i64 inner_after_ = 0;
+    i32 inner_after_ = 0;
 
     {
       do {
