+++ tests/expr/bugs/while2.c
@@ -1,13 +1,13 @@
 #define INTRINSICS_IMPLEMENTATION
 #include "intrinsics.h"
 
-i64 expr(void) {
-i64 result_; {
-    i64 sum_ = 0;
+i32 expr(void) {
+i32 result_; {
+    i32 sum_ = 0;
 
-    i64 count_ = 5;
+    i32 count_ = 5;
 
-    i64 tmp1; {
+    i32 tmp1; {
       while ((count_ > 0)) {
         {
           tmp1 = ((count_ = (count_ - 1)));
