+++ tests/expr/do.c
@@ -1,11 +1,11 @@
 #define INTRINSICS_IMPLEMENTATION
 #include "intrinsics.h"
 
-i64 expr(void) {
-i64 result_; {
-    i64 sum_ = 0;
+i32 expr(void) {
+i32 result_; {
+    i32 sum_ = 0;
 
-    i64 a_ = 0;
+    i32 a_ = 0;
 
     {
       do {
@@ -210,7 +210,7 @@ i64 result_; {
       }
     };
 
-    i64 answer_; {
+    i32 answer_; {
       do {
         {
           {
@@ -237,7 +237,7 @@ i64 result_; {
       }
     };
 
-    i64 tmp29; {
+    i32 tmp29; {
       do {
         {
           ;
