+++ tests/expr/while.c
@@ -1,25 +1,25 @@
 #define INTRINSICS_IMPLEMENTATION
 #include "intrinsics.h"
 
-i64 expr(void) {
-i64 result_; {
-    i64 x_ = 0;
+i32 expr(void) {
+i32 result_; {
+    i32 x_ = 0;
 
-    i64 n_ = 10;
+    i32 n_ = 10;
 
-    i64 sum_ = 0;
+    i32 sum_ = 0;
 
     {
       while ((x_ < n_)) {
         {
           (sum_ = (sum_ + x_));
 
-          i64 tmp4; {
+          i32 tmp4; {
             tmp4 = x_;
 ;
           };
-          i64 tmp5 = 1;
-          i64 tmp3 = tmp4 + tmp5;
+          i32 tmp5 = 1;
+          i32 tmp3 = tmp4 + tmp5;
           (x_ = tmp3);
         };
         continue2_:;
@@ -27,7 +27,7 @@ i64 result_; {
       sum_;
     } break1_:;
 
-    i64 count_ = 10;
+    i32 count_ = 10;
 
     {
       while ((((count_ = (count_ - 1))) > 0)) {
@@ -39,7 +39,7 @@ i64 result_; {
 
     (count_ = 10);
 
-    i64 a_; {
+    i32 a_; {
       while ((((count_ = (count_ - 1))) > 0)) {
         a_ = (sum_ = (sum_ + 1));
         continue9_:;
@@ -55,7 +55,7 @@ i64 result_; {
       }
     };
 
-    i64 b_; {
+    i32 b_; {
       while ((sum_ > 0)) {
         b_ = (sum_ = (sum_ - 1));
         continue11_:;
